chore: improve twenty-ui storybook configuration
