+++ packages/twenty-ui/.storybook/main.ts
@@ -3,17 +3,18 @@ import { StorybookConfig } from '@storybook/react-vite';
 const config: StorybookConfig = {
   stories: ['../src/**/*.stories.@(js|jsx|ts|tsx|mdx)'],
   addons: [
+    '@storybook/addon-links',
     '@storybook/addon-essentials',
+    '@storybook/addon-onboarding',
     '@storybook/addon-interactions',
+    '@storybook/addon-coverage',
     '@storybook/addon-themes',
+    'storybook-addon-cookie',
+    'storybook-addon-pseudo-states',
   ],
   framework: {
     name: '@storybook/react-vite',
-    options: {
-      builder: {
-        viteConfigPath: 'vite.config.ts',
-      },
-    },
+    options: {},
   },
 };
