--- packages/twenty-ui/src/components/Pill/Pill.stories.ts
@@ -1,20 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-
-import { ComponentDecorator } from '../../testing/decorators/ComponentDecorator';
-
-import { Pill } from './Pill';
-
-const meta: Meta<typeof Pill> = {
-  title: 'UI/Display/Pill/Pill',
-  component: Pill,
-};
-
-export default meta;
-type Story = StoryObj<typeof Pill>;
-
-export const Default: Story = {
-  args: {
-    label: 'Soon',
-  },
-  decorators: [ComponentDecorator],
-};
