+++ packages/twenty-ui/src/nx-react.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="@nx/react/typings/cssmodule.d.ts" />
+/// <reference types="@nx/react/typings/image.d.ts" />
+/// <reference types="@nx/react/typings/styled-jsx.d.ts" />
