+++ packages/twenty-ui/tsconfig.spec.json
@@ -9,7 +9,9 @@
     "jest.config.ts",
     "src/**/*.d.ts",
     "src/**/*.spec.ts",
+    "src/**/*.spec.tsx",
     "src/**/*.test.ts",
+    "src/**/*.test.tsx",
     "vite.config.ts"
   ]
 }
